Finally working. We have a concurrent plugin to organise and distribute the work between different theories. It uses Jane Street's Async library. Answers seem correct. Performances seem like before. Architecture and code is rather clean, now (IMHO).
Answers were incorrect, and erratic/inconsistent (e.g. sensitive to the printf calls for stdout printing!, sensitive to native/byte code compilation!),
 when using weak pointers for the hash-consing table (module Weak of the standard lib).
Filed a bug report to the ocaml dev team (apparently there is a memory corruption bug either in Format or in Weak).
Not using Weak at the moment (but rather the usual Hashtbl), until the bug is fixed.