Cleaned up main sequent calculus search function in kernel/theories/prop
- used optional argument to be much more synthetic
- used records to package function arguments together
As a result of this, the search function runs a lot slower than before,
since it now has to package and unpackage records:
{ previous with field = newvalue }
is very nice to write, and makes the code a lot more readable,
but it comes at a small computational price, when compared to the previous style in which functions were given up to 8 or 9 arguments.

Used GADT to distinguish focussed and unfocussed sequents, which is a change that occurs from sequent.ml and requires propagating
a lot of changes throughout Psyche.
I also finally renamed the 2 constructors for the binary sum type as
Case1
Case2
(used to be A and F, for obscure legacy reasons).
That is also a change that propagates through many files.